# swg/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "gopokemon service API",
        "title": "gopokemon service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/user-pokemon/my-pokemon": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Pokemon"
                ],
                "summary": "page user pokemon",
                "responses": {
                    "200": {
                        "description": "json with success = true",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.UserpokemonRes"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "json with error = true",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User Pokemon"
                ],
                "summary": "update user pokemon",
                "parameters": [
                    {
                        "description": "Req Param",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.updateUserpokemonReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "json with success = true",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/model.UserpokemonRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "json with error = true",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.updateUserpokemonReq": {
            "type": "object",
            "required": [
                "nickname",
                "userpokemonId"
            ],
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "userpokemonId": {
                    "type": "integer"
                }
            }
        },
...
        "model.UserpokemonRes": {
            "type": "object",
            "properties": {
                "nickname": {
                    "type": "string"
                },
                "pokemon": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "userpokemonId": {
                    "type": "integer"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "payload": {
                    "type": "object"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payload": {
                    "type": "object"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}